requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.3
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
gunicorn==21.2.0
python-dotenv==1.0.0
dj-database-url==2.1.0
yfinance==0.2.44
drf-yasg==1.21.7